aol_color_rec/top&pants/requirements.txt
numpy
kafka-python
joblib
scikit-learn
opencv-python
pandas
matplotlib
os
sys
json
sqlite3

